Move iio_scan_block API to compat lib
The iio_scan_block API is replaced by the new iio_scan API.
