+++ scan.c
@@ -147,57 +147,6 @@ void iio_scan_context_destroy(struct iio_scan_context *ctx)
 	free(ctx);
 }
 
-struct iio_scan_block {
-	struct iio_scan_context *ctx;
-	struct iio_context_info **info;
-	ssize_t ctx_cnt;
-};
-
-ssize_t iio_scan_block_scan(struct iio_scan_block *blk)
-{
-	iio_context_info_list_free(blk->info);
-	blk->info = NULL;
-	blk->ctx_cnt = iio_scan_context_get_info_list(blk->ctx, &blk->info);
-	return blk->ctx_cnt;
-}
-
-struct iio_context_info *iio_scan_block_get_info(
-		struct iio_scan_block *blk, unsigned int index)
-{
-	if (!blk->info || (ssize_t)index >= blk->ctx_cnt) {
-		errno = EINVAL;
-		return NULL;
-	}
-	return blk->info[index];
-}
-
-struct iio_scan_block *iio_create_scan_block(
-		const char *backend, unsigned int flags)
-{
-	struct iio_scan_block *blk;
-
-	blk = calloc(1, sizeof(*blk));
-	if (!blk) {
-		errno = ENOMEM;
-		return NULL;
-	}
-
-	blk->ctx = iio_create_scan_context(backend, flags);
-	if (!blk->ctx) {
-		free(blk);
-		return NULL;
-	}
-
-	return blk;
-}
-
-void iio_scan_block_destroy(struct iio_scan_block *blk)
-{
-	iio_context_info_list_free(blk->info);
-	iio_scan_context_destroy(blk->ctx);
-	free(blk);
-}
-
 static bool has_backend(const char *backends, const char *backend)
 {
 	return !backends || iio_list_has_elem(backends, backend);
